Move CompilationTaskContext to toolchain, as it doesn't really belong in utils.
